[PTX-22914] Add ValidatePxLicenseSummary
